Add CMakeLists.txt for the entire project
This commit add a CMakeLists allowing to easily
generate native makefile for the tests, exercises
and answers

This commit also update the README.md  and add test harness allowing
to easily run the smoke tests.

Note that currently only python2.7 us supported, to
be able to support python3, a top-level option
named "PYTHON_VERSION_MAJOR" could be introduced
with a default value of 2.